+++ core/src/com/azokh/blackbox/gameselectionscreen/BeamGameSelectionScreen.java
@@ -1,7 +1,7 @@
 package com.azokh.blackbox.gameselectionscreen;
 
-import com.azokh.blackbox.BeamElement;
-import com.azokh.blackbox.ui.Element;
+import com.azokh.blackbox.ui.element.beam.BeamElement;
+import com.azokh.blackbox.ui.element.Element;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Color;
 
@@ -10,8 +10,8 @@ public class BeamGameSelectionScreen implements Element {
     BeamElement beamElement1, beamElement2;
 
     public BeamGameSelectionScreen(int title_delta, int title_height, Color laserColor1, Color laserColor2) {
-        beamElement2 = new BeamElement(Gdx.graphics.getWidth(), 2.0f, Gdx.graphics.getHeight()/3-(int)title_height-65, laserColor2, false, false);
-        beamElement1 = new BeamElement(Gdx.graphics.getWidth(), 2.0f, 2*Gdx.graphics.getHeight()/3-(int)title_height+65, laserColor1, false, true);
+        beamElement2 = new BeamElement(Gdx.graphics.getWidth(), 2.0f, Gdx.graphics.getHeight()/3-title_height-65, laserColor2, false, false);
+        beamElement1 = new BeamElement(Gdx.graphics.getWidth(), 2.0f, 2*Gdx.graphics.getHeight()/3-title_height+65, laserColor1, false, true);
 
     }
